Updated battle_ai_switch_items.c
An update is to Yawn at line 446 because that way the HP check is more specific to if it can come back in. No point switching if HP is 34% but dies on switch in.
My change is to the Switching out during Perish Song AI. If the AI is perish songed and the player has a way to trap the AI, it will preemptively switch out in a way similar to players playing against Perish Song in VGC. It applies to doubles and singles. It may need to be changed to use code previously added to upcoming that I did not have updated to my branch